requirements.txt
# AFlow + ROLL Integration - Requirements
# Python 3.10+

# Core Deep Learning
torch==2.6.0
transformers==4.57.1
peft==0.12.0

# Training & Optimization
numpy==2.2.6
pytorch-lightning==1.9.5

# Experiment Tracking
wandb==0.17.4

# Configuration & Data
pyyaml==6.0.1
tqdm

# Async & HTTP
aiohttp
asyncio

# Optional: For better performance
accelerate>=0.20.0
bitsandbytes>=0.41.0

# Development & Testing (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
